Cleaned up
Moved the Workout database table to models.py and cleans up the WorkoutPage.py file
